# specs/001-governed-url-shortener/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Governed Agentic URL Shortener API
  version: 0.1.0-proposed
  description: >
    PROPOSED contract, not yet approved (Human Gate 4 pending). Traces to
    spec.md FR-101..FR-116 (domain), FR-201..FR-206/FR-501..FR-503
    (orchestration), FR-301..FR-313 (governance), FR-401..FR-406
    (reliability), FR-601..FR-606 (audit/metrics). Backward-compatibility and
    versioning rules: see the "Compatibility and Change Control" section at
    the bottom of this file. Any change to this contract requires impact
    analysis and change approval per FR-306.
servers:
  - url: http://localhost:8000
    description: Local demonstration instance (ADR-0010)
security:
  - {}  # domain endpoints are unauthenticated per spec Exclusions (no end-user auth)
paths:
...
  /workflows:
    post:
      summary: Ingest a requirement, creating a workflow instance (FR-201, FR-206, FR-601)
      operationId: createWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowInstance"
  /workflows/{workflowId}:
    get:
      summary: Inspect workflow state, dependency graph, and decision lineage (FR-201, FR-204, FR-206, FR-503)
      operationId: getWorkflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full workflow state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowInstance"
        "404":
          description: Unknown workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Verified per-role credential (ADR-0006). Never a caller-supplied
        name. Never held by the orchestration/agent process in raw form —
        see ADR-0006's isolation mechanism.
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable error code
        message:
          type: string
        requirement_ref:
          type: string
          description: FR identifier this rejection enforces, for traceability
...
    WorkflowCreateRequest:
      type: object
      required: [requirement]
      properties:
        requirement:
          type: string
        scenario:
          type: string
          enum: [greenfield, brownfield, ambiguous]
          nullable: true
        auto_execute:
          type: boolean
          default: false
          description: >-
            T106, added 2026-09-11 (MINOR, additive, backward-compatible --
            existing callers omitting this field are unaffected). When
            true, seeds a minimal demo pipeline (intake -> classify ->
            conditional branch, FR-202/T104) that the in-process background
            scheduler (src/orchestration/live_scheduler.py) drives
            automatically to completion, with no direct scheduler calls
            required from the caller. External agent adapters remain
            fail-closed regardless of this flag -- see ADR-0006.
    WorkflowStage:
      type: object
      required: [id, name, status, artifact_revision]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [pending, ready, running, succeeded, failed_transient, failed_permanent, awaiting_approval, rejected, skipped]
        attempt_count:
          type: integer
        artifact_revision:
          type: string
        depends_on:
          type: array
          items:
            type: string
          description: stage IDs this stage depends on (the explicit dependency-graph edges, FR-201)
    WorkflowInstance:
      type: object
      required: [id, status, stages, decision_lineage]
      properties:
        id:
          type: string
          description: Correlation identifier (FR-601)
        feature_ref:
          type: string
        scenario:
          type: string
          enum: [greenfield, brownfield, ambiguous, "n/a"]
        status:
          type: string
          enum: [running, awaiting_approval, safe_stopped, completed, rejected]
        stages:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStage"
        decision_lineage:
          type: array
          items:
            type: object
            properties:
              decision_type:
                type: string
              detail:
                type: object
              created_at:
                type: string
                format: date-time
...